feat(build): generate TypeScript declaration file
