+++ .npmignore
@@ -21,3 +21,4 @@ bower.json.ember-try
 package.json.ember-try
 
 /node-tests
+/build
