+++ package.json
@@ -17,7 +17,8 @@
     "lint:js": "eslint ./*.js addon addon-test-support app config lib server test-support tests node-tests",
     "start": "ember serve",
     "test": "mocha node-tests/babel-7",
-    "test:all": "ember try:each"
+    "test:all": "ember try:each",
+    "prepack": "node ./build/write-declaration-file"
   },
   "dependencies": {
     "babel-plugin-debug-macros": "^0.2.0",
